Add product card clicking functionality
Add functionality such that when you click on a product card on the shop page, it takes you to product page. Products on the shop page are imported from the JSON file products.js and are merely used for testing purposes, to setup initial functionality. Eventually, these products will be loaded asynchronously from a server.